+++ src/Pages/signin/Components/SignInForm.tsx
@@ -3,16 +3,19 @@ import styled from "styled-components";
 import Background from "../../../assets/StrokeBackground.png";
 import LoginLogo from "../../../assets/LoginLogo.png";
 //React router
-import {Link} from 'react-router-dom';
+import { Link } from "react-router-dom";
+import { makeStyles } from "@material-ui/core/styles";
+import { TextField } from "@material-ui/core";
+
 //color
-export const colors = {
+const colors = {
   grey: "#D3D3D3",
-  pink:"#EF5DA8",
-  white:"#FFFFFF"
+  pink: "#EF5DA8",
+  white: "#FFFFFF",
 };
 
 //background
-export const StyledContainer = styled.div`
+const StyledContainer = styled.div`
   margin: 0;
   min-height: 100vh;
   display: flex;
@@ -25,7 +28,7 @@ export const StyledContainer = styled.div`
 `;
 
 //loginbox
-export const StyledFormArea = styled.div`
+const StyledFormArea = styled.div`
   position: absolute;
   width: 500px;
   height: 400px;
@@ -37,7 +40,7 @@ export const StyledFormArea = styled.div`
 `;
 
 //loginlogo
-export const Logo = styled.div`
+const Logo = styled.div`
   width: 250px;
   height: 100px;
   background-image: url(${LoginLogo});
@@ -48,8 +51,8 @@ export const Logo = styled.div`
 `;
 
 //loginbutton
-export const StyledFormButton = styled(Link)`
-  background:${colors.pink};
+const StyledFormButton = styled(Link)`
+  background: ${colors.pink};
   box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
   font-size: 18px;
   border-radius: 20px;
@@ -61,19 +64,64 @@ export const StyledFormButton = styled(Link)`
   padding: 16px 32px;
   width: 250px;
   height: 2px;
-  margin-Left: 50px;
-  margin-Right: 50px;
-  margin-Top: 30px;
-  
+  margin-left: 50px;
+  margin-right: 50px;
+  margin-top: 30px;
+
   &:hover {
     background-color: ${colors.pink};
     color: ${colors.grey};
     cursor: pointer;
   }
 `;
+const useStyles = makeStyles(() => ({
+  form: {
+    width: "80%",
+    marginLeft: "10%",
+    marginRight: "10%",
+  },
+}));
 
 const SignInForm: React.FC = () => {
-  return <></>;
+  const classes = useStyles();
+  return (
+    <>
+      <StyledContainer>
+        <StyledFormArea>
+          <Logo />
+          <form className={classes.form} noValidate>
+            {/*Email*/}
+            <TextField
+              variant="standard"
+              margin="normal"
+              required
+              fullWidth
+              id="email"
+              name="email"
+              label="Email"
+              placeholder="Enter Your Email Address"
+              autoComplete="email"
+              autoFocus
+            />
+            {/* Password */}
+            <TextField
+              variant="standard"
+              margin="normal"
+              required
+              fullWidth
+              id="password"
+              name="password"
+              label="Password"
+              placeholder="Enter Your Password"
+              type="password"
+              autoComplete="current-password"
+            />
+            <StyledFormButton to="/home">Login</StyledFormButton>
+          </form>
+        </StyledFormArea>
+      </StyledContainer>
+    </>
+  );
 };
 
 export default SignInForm;
